0.7.1-12
Fixed a problem where an emphasis marked with asterixes on a bullet line marked with
an asterix would be read as a text line. This is not 100% ok, but it will be very
tricky or unlikely to try to mess this up.

Fixed a problem where a reference link followed by a parenthetical expression would
be captured like an inline link with no content.

Added release notes for 0.7.1-12.